fix(appcomposer): move IAM credentials prompt to SAM sync
